Organize the folder structure
