+++ src/app/models/countries/index.ts
@@ -1,0 +1,8 @@
+export interface Name {
+  get fullName(): string;
+}
+
+export interface Birthday {
+  get age(): number;
+  get date(): string;
+}
