+++ src/base/extensions/__tests__/array.test.ts
@@ -1,12 +1,13 @@
-import { describe, expect, test } from "vitest";
-import "@/base/extensions";
+import { describe, test } from "vitest";
+
+import "../array";
 
 const numbers = [1, 2, 3];
 const strings = ["a", "b", "c"];
 const booleans = [true, false, false];
 
 describe("Array constructor extensions", () => {
-  test("zip: same length", () => {
+  test("zip: same length", ({ expect }) => {
     const actual = Array.zip(numbers, strings, booleans);
 
     expect(actual).toStrictEqual([
@@ -20,7 +21,7 @@ describe("Array constructor extensions", () => {
     ["array1", [1, 2], strings, booleans],
     ["array2", numbers, ["a", "b"], booleans],
     ["array3", numbers, strings, [true, false]],
-  ])("zip: %s", ([, numbers, strings, booleans]) => {
+  ])("zip: %s", ([, numbers, strings, booleans], { expect }) => {
     const actual = Array.zip(numbers, strings, booleans);
 
     expect(actual).toStrictEqual([
@@ -31,20 +32,24 @@ describe("Array constructor extensions", () => {
 });
 
 describe("Array.prototype extensions", () => {
+  test.for<[boolean, unknown[]]>([
+    [true, []],
+    [false, numbers],
+  ])("isEmpty: %s", ([expected, input], { expect }) => {
+    const actual = input.isEmpty();
+
+    expect(actual).toBe(expected);
+  });
+
   test.for<[unknown, unknown[]]>([
     [undefined, []],
     [1, numbers],
     ["a", strings],
     [true, booleans],
-  ])("first: %s", ([expected, inputs]) => {
-    expect(inputs.first()).toBe(expected);
-  });
+  ])("first: %s", ([expected, inputs], { expect }) => {
+    const actual = inputs.first();
 
-  test.for<[boolean, unknown[]]>([
-    [true, []],
-    [false, numbers],
-  ])("isEmpty: %s", ([expected, input]) => {
-    expect(input.isEmpty()).toBe(expected);
+    expect(actual).toBe(expected);
   });
 
   test.for<[unknown, unknown[]]>([
@@ -52,7 +57,9 @@ describe("Array.prototype extensions", () => {
     [3, numbers],
     ["c", strings],
     [false, booleans],
-  ])("last: %s", ([expected, inputs]) => {
-    expect(inputs.last()).toBe(expected);
+  ])("last: %s", ([expected, inputs], { expect }) => {
+    const actual = inputs.last();
+
+    expect(actual).toBe(expected);
   });
 });
