--- src/base/index.ts
@@ -1,1 +1,0 @@
-export type * from "@/base/countries";
