+++ src/base/testHelpers/setup.ts
@@ -1,0 +1,11 @@
+import { VueWrapper } from "@vue/test-utils";
+
+declare module "@vue/test-utils" {
+  interface VueWrapper {
+    [Symbol.dispose](): void;
+  }
+}
+
+VueWrapper.prototype[Symbol.dispose] = function () {
+  this.unmount();
+};
