+++ vitest.config.ts
@@ -8,7 +8,7 @@ export default mergeConfig(
     test: {
       environment: "happy-dom",
       globals: true,
-      setupFiles: ["src/base/extensions"],
+      setupFiles: ["src/base/testHelpers/setup", "src/base/extensions"],
     },
   }),
 );
